Rework of encrypt.sh
The proposed changes clean up the script for readability purposes. It also removes the need for gsed/sed as it uses "keybase status" w/ awk. It will list out users you are following and allow you to message that user w/ a corresponding number rather than typing out a name.